refactor(query): add barrel files to desktop and isomorphic subdirectories
